cli/pyproject.toml
[project]
name = "ngen-cli"
version = "0.1.0"
description = "CLI tool for the NGEN platform"
requires-python = ">=3.11"
license = "MIT"
dependencies = [
    "click>=8.0,<9.0",
    "httpx>=0.27,<1.0",
    "pyyaml>=6.0,<7.0",
    "rich>=13.0,<14.0",
]

[project.scripts]
ngen = "ngen_cli.main:cli"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/ngen_cli"]
